fix delete endpoint to stop non owners from deleting the attachment + fixing 3 api endpoints urls to use singular names + refactor the get work by owner query

+++ backend/src/routers/attachment.py
@@ -9,6 +9,8 @@ from assignments.backend.src import queries
 from assignments.backend.src.models import AttachmentFile, AttachmentID, ErrorModel
 from fastapi.responses import StreamingResponse
 
+from enabled.backend.src.users.users import current_active_user
+
 attachment_router = APIRouter(prefix="/attachment")
 
 
@@ -28,11 +30,17 @@ async def get_attachment(attachment_id: UUID,
 
 
 @attachment_router.delete("/{attachment_id}/delete",
-                          responses={404: {"model": ErrorModel}})
+                          responses={404: {"model": ErrorModel},
+                                     403: {"model": ErrorModel}})
 async def delete_attachment(attachment_id: UUID,
+                            user=Depends(current_active_user),
                             client=Depends(get_client)) -> AttachmentID:
-    result = await queries.delete_attachment(client, attachment_id=attachment_id)
-    if result:
-        return result
-    else:
+    # TODO: find a better way to make sure the attachment exists
+    owner = await queries.get_attachment_owner(client, attachment_id=attachment_id)
+    if not owner:
         raise HTTPException(status_code=404, detail="Attachment NOT found")
+    elif not owner.id == user.id:
+        raise HTTPException(status_code=403, detail="Only attachment owner can delete it")
+    else:
+        return await queries.delete_attachment(client, attachment_id=attachment_id)
+
